backend/requirements.txt
dotenv
flask
flask-mysqldb
flask-cors
flask-sqlalchemy
psycopg2-binary
pillow
opencv-python
numpy
joblib